api-scraper: do not log nan values
At this point, we have a good handle on which columns are nullable from the logs and BigQuery dump.

This updates `_log_invariant_violations` to only log cases where multiple non-null values appear for a field expected to be static. It will no longer log cases where the field value is nan.

Goal: log cleanup.